requirements.txt
yfinance>=0.2.36
akshare>=1.14
pandas>=2.0
pandas-datareader>=0.10
numpy>=1.24
matplotlib>=3.7
fastapi>=0.110
uvicorn>=0.29
PyJWT>=2.8
python-multipart>=0.0.9
